Match kernel version when installing kernel module with yum
Specify the running kernel version when installing the kernel module
with yum.  Otherwise yum will pick the largest release number of the
kmods in the yum repo which may not match the current kernel version on
the target host, which causes the yum install of the kmod to fail.

We also need the version number in order to specify the release
component of the rpm package.  So first install the base package and
then query that package to retrieve the installed package version field.