Refactor client dkms installation tasks
Remove the odd with_dkms "flag" and select dkms installs with a new
value of the client_install_method variable.  Move the dkms tasks to a
separate file and remove the extra when conditionals.

Remove the unneeded task to add the epel repo during the client install.
Remove a spurious when:False condition left over during a debugging
session.