Document password variables
Document the password variables that need to be set, either as playbook,
group variables, or extra variables.

Remove the kludgey shell script which prompted for these passwords and
created encrypted group variables, since the location and encryption
methods are up to the playbook author, and we need less scripts to
maintain.