Adopt per cell inventory directory layout
Adopt a directory layout for the inventories as recommended by the
ansible documentation. In this scheme, we create a subdirectory per cell
with a directory structure for the inventory and variables files.

Provide one example inventory for the example.com cell. This example
includes an empty directory which is git-ignored to save vault encrypted
passwords.

Move the example ansible.cfg out back to the top level directory of
the project.

Fix the tasks which did a "split" of the partition id string, since we
keep that info in a structured yaml file now instead of the hosts ini
file.