Create the CellServDB once and then distribute
Create the CellServDB and ThisCell configuration files from the database
host facts once, at the beginning of the server role, and then
distribute those as ansible managed files to the afs servers and the afs
clients.

Before this change, we would attempt to create the CellServDB on the
clients using a template and the database host facts. However, if we are
running a play for just the clients, the database facts are not
available. Instead, just create the CellServDB file when deploying the
servers and deploy the same contents everywhere. This has the advantage
of ensuring we have a consistent CellServDB file on all the remotes, and
removes some duplication of template files.

The generated CellServDB is kept in the "files" directory in the same
path as the inventory hosts file.