Fail early when the admin password is undefined
We do not have a default value for the admin password, so just fail
early with a better error message if the password was not set in the
inventory or extra (-e command line option) variables.

It could be possible to prompt for the password with the 'pause' task,
but that is a bit complicated, so defer that for now. (Note,
'var_prompt' does not allow use to override with inventory defaults,
only extra variables from the command line.)

Also, user a random string for the default value of the kerberos master
database value. This is sufficient and convenient for test deployments.