Add role prefix to initial cell setup variables
Prefix the variables that define the initial volumes, users, and groups
with the adopted 'afs_' prefix.  Set defaults to be empty (None) to
skip if not defined in the group variables.

Rename 'dir' to 'mtpt' in the afs_volumes to stress this is a mount
point, and use the volume name for the mount point name by default.

Make afs_users an list of dictionaries so we can add other per user
attributes in the future, and to make the structure consitent with the
afs_volumes and afs_groups variables.

Document these cell role variables and provide examples in the readme.