Add a role variable to set the initial user password
The afs_user_password provides the initial Kerberos password for users
created by the cell role, if any. This variable is undefined by default.
Provide a value as an extra variable (-e command line option), or in a
group variable, preferably encrypted with ansible-vault.

The cell role will fail early with a warning message if users are listed
in afs_users but afs_user_password is not defined, even if the
principals already exist in the kdc.