Move afs service key creation to server role
The afs service key creation really belongs to the openafs server role,
not the kdc role.  Doing this in the kdc role limits us to one cell per
realm.

Move the creation of the afs service key to the beginning of the server
role. Run it once, on the first server in the group, since it needs to
be done once for the group. Use the regular kadmin command to add the
afs service principal. kadmin.local is now used only to create the
initial admin principal.