Use ansible_pkg_mgr to select installation tasks
Use the built-in variable ansible_pkg_mgr to select the installation
tasks based for a given remote system instead of trying to select by os
name.

Change the server and client "install method" variable to select the
package manager installation or a `make install` installation to
facilitate development and testing.