Check firewalld service status instead of starting
The firewalld immediate option fails when filewalld is not running, so
we were always starting it first.

Instead of always starting firewalld, just check to see if it is already
running. If it is running, apply the changes "immediately", otherwise
enable the changes for the next time the firewalld service is started.