Add rsync install method
Add a new install method to replace the old 'make-install'.  The new
method is called "rsync" and can be used to install a binary
distribution, located on the remote host.  Attempt to build this
distribution from source if it is not present, for the client and the
server roles.

In the future we may support other methods of retrieving a precompiled
binary distribution for the target arch.

This commit provides rsync install support for redhat. Other platforms
may be added in the future.