Avoid unneccessary bosserver restarts
The bosserver removes the BosConfig.new file on a restart so we cannot
use that file in ansible to detect a configuration change. As a kind of
work-around, stage our configuration in the separate file called
BosConfig.ansible to detect the state change, and in a separate task
copy that file to BosServer.new to trigger the handler to restart the
bosserver.