Use akeyconvert to set the service key
Instead of the clumsy `asetkey` interface, which requires extracting the
kvno and key types, use the nice, new `akeyconvert` tool to import the
Kerberos service key. If `akeyconvert` is not available, assume this is
running a 1.6.x version that can read the rxkad.keytab directly.

`akeyconvert` usage was suggested by Ben Kaduk and the OpenAFS Workshop
2019 in Pittsburgh.