Fix ansible bare variable warnings
Squelch the new ansible bare variable conditional warnings.

The afs_selinux variable is a boolean anyway, so thhat warning seems
bogus anyway, but us the |bool filter to appease recrent versions of
ansible.

The other conditionals are not necessary if we just make the defaults
to be empty lists, since we use them as items.