openafs_server: wait for quorums
Improve the flow by waiting for the prdb and vldb quorums to be
established, plus 15 seconds for the first beacons, before attempting to
write to the databases.  This is functionally the same, but looks better
from an ui perspective than always failing on the first pts createuser
command. We still retry the pts and vos commands just in case quorum
drop.

The server logs are not helpful to know when quorum is established, at
least in current OpenAFS releases, so check the output of `udebug` to
see when a sync site has been elected.  It's all very jangy, and would
be better as an ansible module.

Since these tasks all need to be done on the "first" server, move
them to a new file called "cell-bootstrap.yaml".