Generate random passwords when not defined
When the passwords are not defined as hostvars in the inventory or extra
vars on the command line, instead of failing with a message, generate
random passwords and save the values in files on the controller.  This
intended to be a convenience when setting up cells for testing and we do
not want to bother with ansible-vault for ephemeral test passwords.

This is done for the admin user password, the kdc master password, and
the initial test user password.  The passwords are saved in a set of
hidden files in the users home directory. The files are not namespaced
by cell name.

If you want different passwords per cell, or have existing passwords
which must be used, please define them as hostvars, preferably protected
by vault.