Fix rsync to not remove symlinked directories
If a target directory is a symlink and the source directory is not,
rsync will remove the symlink and create a "real" directory.  This can
cause problems (e.g. on centos /lib is a symlink to /usr/lib).

Rsync has an option (-K) that treats any target symlinked directory as a
real directory.