skip creation of /afs mountpoint if already mounted
If the "/afs" mountpoint is already created, it's okay to rerun the
task.  However, if the "/afs" mountpoint is already _mounted_
(because the cache manager is already running due to a reboot or a playbook
rerun), creation of the "/afs" mounpoint will fail.

Instead, add a task to detect if the mountpoint is already present and
mounted.  If so, skip the task that creates the mountpoint.