Avoid hardcoded /afs mountpoint
Introduce a new variable 'afs_mountpoint' and assign default value /afs
for the roles that need it.

In addition, base the value of afs_cacheinfo_mount on afs_mountpoint to
remove that hardcoded value as well.

The systemd unit files still contain hardcoded 'umount /afs' when
stopping the service - I didn't know how to fix that, so I just added
comments.