krbclient: Add host variables for krb5.conf values
The "afs_kdcs" inventory group is not present when the play is just
setting up clients, so first check host variables "afs_kdc_servers" and
"afs_kadmin_server" to set the kerberos hostnames in the krb5.conf file.
If the variables are not available, then fall back to the "afs_kdc"
inventory group, to be compatible with old playbooks.

Finally, if the variable are not set and the group is not defined, just
leave the krb5.conf realm stanza empty, which is fine when we have DNS
SRV records for kerberos.