Default selinux mode to permissive on CentOS 8
SElinux enforcing mode currently breaks the servers on CentOS 8.

For now set the default mode on CentOS 8 to 'permissive'. The selinux
mode default is now platform dependent. The platform default mode is set
with the constant _afs_selinux_mode, which is set in the platform
specific vars file.

Override this default value by setting the 'afs_selinux_mode' variable
in the inventory or on the command line with the -e option.