Bump version in yum repo url to 1.8.5
Bump the default version number to be installed to 1.8.5, which is
the current latest available in the public SNA yum repo. (In the future
we may have a repo tree organized by platform, so the latest is
installed, and updated to, by default.)

Use the '%releasever' variable instead of hard-coding 'el7' in the
yum repo url.