tests: Initial playbooks test harness
Start of a test framework based on unittest to spin up virtual machines
and run our playbooks against them.  Runs the build* and testcell
playbooks on centos7, centos8, and debian10.  Logs the ansible output to
a file and optionally to the console.

Requires pre-existing template virtual machines to be cloned and
sysprepped. The templates should be updated to have the most recent
kernel levels so the kernel-devel packages can be installed during the
plays.