client: Check /proc/modules for openafs kernel module
Instead of running lsmod, check the contents of /proc/modules to
determine if the openafs module is loaded. Check before running modprobe
to avoid an false 'change', and check after modprobe just to be sure the
module is loaded before continuing.

This change also fixes an error when lsmod is not found in the path.