Create libvirt snapshots after each stage
Optionally create libvirt snapshots after each stage so we can quickly
rollback hosts to the most recently completed stage with 'virsh
snapshot-revert <name> --current'

Update the vm.py helper to remove the snapshots in the test fixtures.

To enable the snapshots, add the libvirt_domain host variable in your
inventory for the hosts to be snapshotted.