Add openafs_install module exclude option
Add an option to exclude directories and files from the installation.
This can be used to avoid installing client files on a server, or server
files on a client, or other obsolete files. The exclude option can be a
single string or a list of file pattern strings (glob notation).

Also, create the install log directory if it does not already exist.