Add from-source installation method
Add the 'from-source' installation method for clients and servers, which
builds and installs the OpenAFS binaries with the new custom
openafs_build and openafs_install modules.

In order to support arbitrary configure options, change the installation
paths to be packaging specific, instead of OS specific. The custom
openafs_install module will set the installation paths as facts.

The old 'rsync' build and install method is deprecated and will be
removed soon.