Add OpenAFS installation local facts
Currently, we assume the paths to OpenAFS commands depend on the OS
platform, which is only true if we are using distribution provided
binaries.

Instead of assuming the paths to commands depend on the platform type,
query the paths installation and save them as Ansible local paths.
Convert the roles to use the local facts for paths to OpenAFS commands
and installation and configuration directories.

The local facts are saved in the json file
/etc/ansible/facts.d/openafs.fact on the remote systems where OpenAFS is
installed by these roles. This fact file is automatically read when
Ansible gathers facts on subsequent plays.

Users can override the installation paths with host variables, if from
some reason you want the roles to use some alternative you've installed
outside of the roles.

This change is a prereq to support binary distribution installations.