Managed install method
Refactor the clients and servers to reintroduce "install methods" to
determine how OpenAFS is to be installed. The default install method is
"managed" which uses the distribution's package manager to install
OpenAFS.  The install method is the same for the client and server roles
on a given remote node, but can vary from node to node.