Add openafs_add_keys module to set the service keys
Currently, we set the service keys on each server by trying to run
kadmin on each server, which is confusing and error prone, and means the
kadmin server needs to be running and reachable. Also, we use the newer
akeyconvert utility to import the keys instead of asetkey, but akeyconvert
is not available on older versions, or may not be packaged on older platforms.

Instead, require service key to be created and exported to a keytab file
once, before any server roles are run. Provide a new module called
openafs_add_keys to idempotently import the service keys with asetkey
utility, and cope with the key version numbers, enctypes, and extended
key support.

The openafs_add_keys module detects if the asetkey program supports
extended keys by running checking the asetkey usage output. The module
will also decode the keytab to find the kvno and enctypes needed for the
asetkey arguments.  The keytab is decoded directly to avoid a klist
dependency on the server.

If extended key support is not found in the installed asetkey program
and the keytab does not contain any DES keys, then just assume the
server has rxkad-k5 support (openafs 1.6.5+ .. 1.6.x) and will read the
rxkad.keytab directly.

Update the readme to describe how a keytab is a prereq and remove no
longer needed kerberos related variables (path to kadmin, kdc host name,
enctypes, etc.).

Add debian 9 back to the molecule tests since we no longer depend on
akeyconvert.