Add afs_csdb variable to setup CellServDB files
Currently, the client and server roles create and update the CellServDB
file from  gathered facts of hosts in the afs_databases group.  This
means the roles will not work unless the database servers are in the
play.

Instead, support a new variable called afs_csdb which contains a list of
host addresses and names. The afs_csdb is required for server roles and
is optional for client roles.  The afs_csdb values may be provided as
host variables in the inventory or as an external yaml formatted file
which is loaded into the afs_csdb variable by the client and server
roles.

Provide a new task in the common role which can be used to generate a
csdb.yaml file from the hosts in the afs_databases group. This can be
used in a playbook before installing OpenAFS on the servers and the
clients.

While here, clean up the tasks that setup the client and server
configuration files.