Remove driver_libvirt_uri from the molecule cookiecutter template
This setting is confusing since it looks like a connection settting from
the ansible controller, but it is really to setup the
LIBVIRT_DEFAULT_URI environment variable for the virt_up module which
runs on the hypervisor.

Instead, remove this variable from the cookiecutter create template (it
already missing in the destroy template).  The LIBVIRT_DEFAULT_URI
environment variable must be correctly configured on the hypervisor
(which may be remote) as a prereq.