Add afs_install_archive variable
Make it easier to specify the installation files.

Remove the lookup of the files to be installed and add the
`afs_install_archive` variable to explicitly give the fully qualified
path on the controller. This has the added benefit in that the `bdist`
method is the same for modern and transarc style builds.

For `bdist` installs, this file contains the results of the `make
install DESTDIR=...` or `make dest`.  For `packages` installs, this file
contains the set of rpm packages files.

Upload this single file instead of a set of rpms in a directory. This
should help avoid mixing different package versions in a given
directory.

Remove the various rpm and bdist variables that are no longer needed.

Change the openafs_devel build scenario to put all of the build results
in a signal flat directory. The path is set in with the cookiecutter
'builds' parameter, default: ~/.cache/ansible-openafs/builds. (Note,
this is intentionally not a role variable.)

Update the molecule scenarios to lookup the most recent archive created
by the openafs_devel role tests. For now, break our rule about using
conditionals in the molecule scenarios.

Remove the separate `bdist-transarc` scenarios, which are no longer
needed since the roles no longer need to know the bdist style to lookup
the path of the files to upload.

Use env vars in the molecule tests to specify the archive file
to be installed and the git repo and ref for builds.