Avoid raising SIGXCPU to retry vl registration
Unfortunately, a side effect of using SIGXCPU to force the vl
registration retry is the fileserver dumps the callbacks and hosts
tables, which is not only extremely heavyweight, is also buggy and can
cause the fileserver to crash if done during startup.

Disable the SIGXCPU by default. Set `afs_registration_retry_signal` to
yes to enable the periodic signal to avoid the fileserver's built in 5
minute retry delay.