Add 'sdist' install method
Add the 'sdist' (source distribution) install method to build and then
install the binaries from the source distribution tarballs generated by
"make dist" and published upstream.

The src and doc tarballs, along with any other files, such as md5 files,
ChangeLog, and RELEASE notes should be bundled in a single compresse
file (tar.gz) for distribution to the remote nodes. Set the
afs_install_archive to specify the path to this file on the controller.