Add afs_topdir variable to customize top build directory path
Add the afs_topdir variable to specify the path to the top build
directory for the scm and sdist install methods. The default continues
to be '~/openafs' on the remote node.

Add the topdir cookiecutter parameter to customize the afs_topdir in the
molucule tests.

Limit the spans of the {% raw %} and {% endraw %} tags in the molecule
scenario cookiecutter.