Start fileserver after db servers
When the fileserver starts, it tries to contact the vl server to set the
UUID and addresses. When it fails, it retries later, but waits 5
minutes, which is quite a long time.  To avoid this delay, check for the
presence of BosConfig, and if not found, defer the fileserver startup
until after the database quorom is detected.

The 'afs_server_cold_start' host variable can be set to 'no' to disable
deferring the fileserver startup, or set to 'yes' to force the cold
start tasks. The default is to check for the presence of BosConfig.