Add afs-scenario to create molecule files
Replace the hard to maintain cookiecutter with a tool called
afs-scenario to create templated molecule files.  The afs-scenario tool
has three subcommands:

create - Create the initial files for a new molecule scenario using the
built-in cookiecutter templates. The developer should edit and commit
the files as needed after running init. Instead of creating a
molecule.yml file, create a molecule.yml.j2 template to support
different drivers.

setup - Setup a driver specification file (driver.yml) which
indicates the driver name and options to be used on the test system.
This will create a yaml file which can be stored on the developer's
system.

init - Create the molecule.yml files from the template and the
driver.yml file. Regular molecule commands can be run after the init
command has been run to generate the molecule.yml files.