Squash yaml and ansible lint errors
Fix makefile rules and role meta data files to be able to run the lint
checks again.  Fix minor projects in the role files to fix new lint
errors.

Suppress the 'risky-file-permissions' ansible lint rule for now, since
we need to explicitly specify all of the mode bits to suppress this
rule.