Add afs_csdb_file default value
Set the default value of afs_csdb_file to ~/.ansible-openafs/<cellname>/csdb.yml
so we don't need to define in the playbooks unless we want to use a
different path or name.

Update the documention and give examples to better explain afs_csdb and
afs_csdb_file.

Update the example cell playbook to import generate_csdb only when
afs_csdb is not defined.

Fix generate_csdb for the case when afs_csdb_file is just a file
name, e.g. afs_csdb_file=csdb.yaml