Update kadmin acls when adding/removing principals
Instead of using a template to create a fixed number of acl entries, add
an optional acl param to the openafs_principal module to also add kadmin
acls when creating a new principal. Also remove acls when a principal is
deleted.

This change disentangles the krbserver role from the afs roles, since we
no longer need to know the afs admin principals to render the template.