Extract docs from modules
Update the embedded docstrings so docs are cleanly generated.

Use the ansible-doc-extractor to extract the restructured text from the
embedded documentation strings in the modules.

For this to work, the project parent directories must be named correctly
to mimic the ansible collection paths:

  $ mkdir -p ansible_collections/openafs_contrib
  $ cd ansible_collections/openafs_contrib
  $ git clone https://github.com/openafs-contrib/ansible-openafs openafs
  $ cd openafs
  $ make init
  $ . .venv/bin/activate
  $ make docs

The generated docs are included in the git repo so readthedocs can
render the html.