Add NetInfo/NetRestrict file support
Add support for the server and client NetInfo and NetRestrict
configuration files using a set of host variables.

This commits adds the host variables which can be defined in your
inventory:

  afs_server_netinfo       sets the server NetInfo
  afs_server_netrestrict   sets the server NetRestrict
  afs_client_netinfo       sets the client NetInfo
  afs_client_netrestrict   sets the client NetRestrict

The contents of these variables are copied to the corresponding
configuration file. A single value can be specified as a string, or
multiple entries may be specified as a list of strings.

When a variable is empty or undefined, the corresponding file
will be removed.

The server will be restarted after changing the values of the server
NetInfo and/or NetRestrict files.

Examples variables:

  afs_server_netinfo:
    - 192.168.1.123
    - f 10.1.1.321

  afs_server_netrestrict: 196.168.0.0/16