Support default client csdb again
When we added a default value for the afs_csdb_file, this masked the
ability of the client role to define a default csdb when afs_csdb is not
defined.

Use an empty afs_csdb_file variable to indicate "not set" so the client
will use a default csdb when afs_csdb is not defined and afs_csdb_file
is empty.

Update the client role molecule files to use a default csdb again, which
is just an empty csdb with the cellname.