modules: Determine make target from configure options
By default, run 'make dest' if the given configure options contain
--enable-transarc-paths, otherwise run 'make install'.  Also, by
default, add the _nolibafs target suffix if the configure options
contain the --disable-kernel-module option.

This makes it possible for the caller to specify the configure options
and have the make do the right thing.

The caller can specify an explicit make target if desired, so this
change is compatible with prior versions.