Add afs_configure_options variable
Add the afs_configure_options variable to control the configure options
used when the afs_install_method is "source" or "sdist".

The afs_configure_options can be a string containing of the configure
options, e.g.,  "--enable-debug --enable-checking", or a dictionary.  The
string format is easier to use when specifying the configure options
with extra vars or environment variables. The dictionary format is
intended for use in playbooks and inventories.