modules: Do not create a temp dir and file in store_facts
In an attempt to make openafs_store_facts idempotent, the facts json are
first written to a temp file, then the sha1 is compared to the
openafs.fact json file in the ansible facts directory.  Unfortunately,
this does not work.

Firstly, the module puts the temporary file in our conventional
temporary directory, which is used by the roles (/tmp/ansible-openafs),
and creates that directory as root if it does not exist. Since we run
store-facts as root (with become), the created directory is owned by
root, and so is not writable by the regular user, if it happens to not
have been created by the roles before invoking store-file.  As it turns
out, we don't even need an extra directory, since we are updating just
one file.

Secondly, and more importantly, it seems the json written can be
different even if the values of the facts did not change, since the keys
can be shuffled each time the file is dumped (even when using
OrderedDicts).

So remove the writing of the facts to the temporary file, and go back to
always writing once to the openafs.fact json file.  This means for now
we will always show a change again, but at least the module will be
usable again.