Add afs_is_server and afs_is_client variables
When installing from source (and therefore building), we need to know if
the node is both a client and a server, since we need to build the
common usespace libraries once.  The client and server roles can be
imported in any order, so the only way to know is with an inventory
variable or inventory group name.

Add the afs_is_server and afs_is_client variables in the common role.
Check the conventional group names to set these by default.  Playbook
authors may set these variables in their inventory if they have a
different group naming convention.

Verify the afs_is_server is true at the beginning of the server role,
and verify the afs_is_client is true at the beginning of the client
role.

If the node is both a client and a server, at the beginning of the
client role, verify the server role was already imported by checking the
local facts. We want to be sure the server is always installed first,
because we should build the userspace binaries once, and it makes sense
for the servers to be installed first so the root volume exist before
starting a client.