Support rebuilding when installing from source
Rework the 'source' and 'sdist' install methods to remove duplication
and support rebuilding and reinstalling when the source changed since
the previous playbook run.

Unfortunately, when installing from source, we cannot feasibly manage
the server and client installations independently when a node is both a
client and a server. When a node is both a server and a client, build
the userspace and kernel module when the server role is imported, and
require server role import before the client role import.  This way we
only build the binaries once and avoid mixing binaries from different
builds on a given node.

When the node is a server only, build and install just the userspace
binaries when the server role is imported. There is not need to build
the kernel module in that case, which is faster and avoids the
requirement for kernel headers on the server.

The sdist and source install methods now include a common
build-and-install.yaml with has common build and install steps. The
tasks to checkout files from git or to upload and untar a source
distribution file has been moved to the 'openafs_common' role, since
these are the same for the client and the server.

When rebuilding, be sure to shutdown the server and/or client services
so we can reinstall.

Avoid rebuilding and reinstalling when the checkout has not resulted in
a changed.

If rebuilding from a git checkout, always do a git clean to remove build
artifacts before rebuilding. This will be made optional later.