Add afs_always_build and afs_clean_build host variables
Add host variables for the 'sdist' and 'source' install methods (that is
when installing from source).

When afs_always_build is true, the build tasks to not be skipped even when
the checkout task did not change. The default value is false.

When afs_clean_build is true, the build module will perform a 'git clean'
if a .git directory is found, otherwise the module will run 'make clean',
so a full build will be done. The default value is true.