build: Do not check git to skip build
Remove the code to try to detect if the git checkout changed since the
last build and if so skip the build.

The roles have been changed to skip the build if the checkout is
unchanged, for the git or non-git checkouts, so this extra complexity is
not needed, and actually impedes the 'afs_always_build' feature.