build: Reinstate skipping regen.sh if configure already exists
The configure script is already provided when building from a
source distribution.

If building from git, configure will be generated the first time the
build task is run, and when rebuilding will be reused unless 'clean' is
true, in which case it will be removed and then regenerated.  This means
if you are changing branches or changing m4 between runs, make sure
'clean' is true to regenerate configure.