Add afs_transarc_build variable
Add an option to add --enable-transarc-paths to the configure option
when the install method is 'sdist' or 'source'.

This is just a convenience so we can let the roles pick default
configure options but still install in the old fashioned transarc style,
e.g., /usr/afs/bin, /usr/afs/etc, /usr/vice/etc.