Rebuild if last build did not complete
When installing from source, use a local fact to remember if the last
build successfully completed, and if not automatically try to build
again the next time the role is imported.  Before this change, we would
rebuild only if the source checkout changed or the always build flag was
set.