test: Update makefile test targets
Update the helper makefile for the client and server roles to run tests
by OS or by scenario.  Add some env vars to help customize the test
harness.

Fix the server test-debian10 target so it actually tests debian10
instead of centos8.

When running all tests, skip 'packages' install method tests on debian
until we have an automated way to create the debs.  (This can still be
tested manually with `make mol` or `molecule`.)