bdist install method is for transarc paths
Binary distribution installation is supported legacy Transarc paths
mode.  It doesn't make sense to support modern paths with a bare
tarball.  Modern paths require packaging (such as deb packages) or
building and installing from source code.

Remove "bdist_modern" from the molecule tests and update the
documentation to mention bdist is for Transarc style paths.

To create a bdist tarball, configure with --enable-transarc-paths. On
linux, also specify --with-linux-kernel-packaging, and build with `make
dest`.  The dest directory should be saved as a tarball on the
controller.