Run the bosserver in restricted mode by default
Fortunately, we use the BosServer.new file to setup the bnodes, so bos
create is not needed by the ansible roles, and we can easily run the
bosserver in this more secure mode by default.

By default, run the bosserver in restricted mode. Add and document the
new bosserver_restricted_mode variable (a boolean) to specify if the
bosserver is to be placed in restricted mode or not. The bosserver and
any running processes under the bosserver are restarted to change the
mode.