client: Fix dkms status check
Fix the incorrect regex pattern to verify that the openafs module was
installed via dkms, when installing from packages (either with rpm files
or a yum repo).  Currently, we fail even though the module was installed
with dkms.

When a dkms module is installed, the dkms status is in the form:

   <module>/<version>, <kernel-version>, <arch>: installed

Thanks to Ralf Brunckhorst for pointing out this bug.