tests: Run molecule tests with pytest
Add a pytest script to run molecule tests for roles and playbooks.  Use
pytest to parameterize tests for each combination of platform and
scenario.

Replace the makefile recipes that run molecule and replace those with a
simple recipe to run pytest. Remove all the targets to select tests.
Tests can be selected with the pytest -k option now.