tests: Avoid CellServDB download when making SRPM
Use the in-tree CellServDB file to build the source RPM instead
of trying to download the file from grand.central.org, when running
the openafs_devel tests.

Do this by extracting the makesrpm.pl script from the src tar ball and
running it directly. Since we get the script from the src tar ball,
there is no longer a need to run 'configure' to make the SRPM.