devel: Improve kernel headers installation
The kernel devel packages must match the target kernel version in order
to build a usable kernel module. Unfortunately, the installation of the
kernel headers (kernel-devel package on redhat) can fail when the kernel
version is too old on the target node (unless extra work is done to find
the old package versions.)

When we cannot install the kernel headers for the running kernel,
optionally update the kernel and reboot so we install the most recent
kernel and the matching kernel headers.

This commits adds 2 new host variables to the openafs_devel role:

  afs_devel_kernel_headers - install the kernel-devel packages
  afs_devel_allow_kernel_update - allow kernel update and reboot

Use rpm-query and dpkg-query to check if the kernel headers are
installed, since the package manager can be slow.  Remove the "update
kernel" tasks in the molecule test playbooks, and use the openafs_devel
role to install the kernel headers.