Tailable logs in openafs_build module
Instead of gathering all of the long running command output into a large
string and dumping it after the commands run, use Popen to run the build
commands and redirect the stdout and stderr to log files. This make it
possible to watch the progress and should use far less memory.