tests: Move molecule driver config to site-local files
Remove the hardcoded driver section from the molecule files so we can
create site-local driver configuration in site-local molecule base
configuration files.

The default path for the base config is "~/.config/molecule/config.yml".
This can be overridden with the molecule --base-config (-c) option.

For example, to use vagrant with a libvirt provider as the molecule
driver, create a config file "~/.config/molecule/config.yml" with the
contents:

    ---
    driver:
      name: vagrant
      provider:
        # Can be any supported provider (virtualbox, parallels, libvirt, etc)
        name: libvirt