modules: Add module_utils Logger class
Replace duplicated logging code in each module with a new module_utils
Logger class, which is just a wrapper over the standard Python syslog
module.

We are not using all of the complex features of the logging framwork,
and using the syslog module avoids the need for platform specific
logging handler setup, e.g. solaris vs linux.