Fix kdc hostnames in krb5.conf
Write the kdc's hostname in the krb5.conf on the kdc instead of the
ansible inventory name (which may not match the actual hostname).

Do not clobber the kdc's krb5.conf file when the client role is
imported.

For now, the openafs_kdcserver role assumes there is only one kdc. This
can be workedaound in a playbook, or we can add more variables in the
future.