tests: Fix path to hosts file on solaris in test preamble
Fix the path to the hosts file on solaris when setting up the hosts
file before running tests. This is not part of the roles or playbooks,
but rather a workaround that runs before running the molecule tests.

Add the AFS_PREP_REWRITE_HOSTS option to enable the workaround to
rewrite the hosts files on the tests instances. This can be useful when
a proper nameserver has not be setup for the test instances.