server: Limit changing the hosts file to Debian
Debian systems add a loopback entry to the /etc/hosts file as a
workaround which causes the hostname to resolve to a loopback address.
This will cause the bosserver to fail on startup.  Remove the offending
loopback entry, but limit to Debian based systems.

For all platforms, add a hostname to address resolution check before
startup. Verify we are able to resolve the hostname and that the
resolved address is not a loopback.